provide_session keep return type (#9787)

+++ airflow/api_connexion/schemas/pool_schema.py
@@ -59,7 +59,7 @@ class PoolSchema(SQLAlchemySchema):
         return obj.queued_slots()
 
     @staticmethod
-    def get_open_slots(obj: Pool) -> int:
+    def get_open_slots(obj: Pool) -> float:
         """
         Returns the open slots of the pool.
         """
